Adjust namespaces of moved controls
